feat: enhance recipe management with nutritionist integration and detailed templates

+++ services/recipes-service/src/main/java/pe/edu/upc/recipes_service/recipes/interfaces/rest/transform/FavoriteRecipeResourceFromEntityAssembler.java
@@ -12,7 +12,7 @@ public class FavoriteRecipeResourceFromEntityAssembler {
         return new FavoriteRecipeResource(
                 favoriteRecipe.getId(),
                 favoriteRecipe.getUserId(),
-                favoriteRecipe.getRecipe().getId(),
+                favoriteRecipe.getRecipe().getId().intValue(),
                 favoriteRecipe.getRecipe().getName()
         );
     }
