+++ services/recipes-service/src/main/java/pe/edu/upc/recipes_service/recipes/interfaces/rest/transform/IngredientResourceFromEntityAssembler.java
@@ -10,7 +10,7 @@ public class IngredientResourceFromEntityAssembler {
 
     public static IngredientResource toResourceFromEntity(Ingredient ingredient) {
         return new IngredientResource(
-                ingredient.getId(),
+                ingredient.getId().intValue(),
                 ingredient.getName(),
                 ingredient.getCalories(),
                 ingredient.getProteins(),
